requirements.txt
Django>=3.0.1,<3.1.0
flake8>=3.6.0,<3.7.0
psycopg2-binary>=2.8.3,<2.9.0
djangorestframework>=3.11.0,<3.12.0
python-dateutil>=2.8.1,<2.9.0
celery>=4.2.1, <4.3.0
redis>=2.10.6, <2.11.0
gunicorn>=20.0.4,<20.1.0
